Rearranging model classes
